requirements.txt
kagglehub
numpy
Pillow
PyYAML
pytest
scikit-learn